+++ src/components/RegisterForm.jsx
@@ -16,20 +16,23 @@ const RegisterForm = () => {
   const [errors, setErrors] = useState({});
   const [isEmailVerified, setIsEmailVerified] = useState(false); // 이메일 인증 상태
   const [isModalOpen, setIsModalOpen] = useState(false); // 이메일 인증 모달 상태
-  const [verificationCode, setVerificationCode] = useState(""); // 인증 코드 입력값 
+  const [verificationCode, setVerificationCode] = useState(""); // 인증 코드 입력값
   const schema = Joi.object({
     username: Joi.string().min(1).max(30).required().messages({
       "string.empty": "",
       "string.max": "30글자 이하로 입력해주세요.",
     }),
-    email: Joi.string().email({ tlds: { allow: false } }).required().messages({
-      "string.empty": "",
-      "string.email": "유효한 이메일이 아닙니다.",
-    }),
+    email: Joi.string()
+      .email({ tlds: { allow: false } })
+      .required()
+      .messages({
+        "string.empty": "",
+        "string.email": "유효한 이메일이 아닙니다.",
+      }),
     password: Joi.string()
       .required()
       .pattern(
-        /^(?=.*[A-Za-z])(?=.*\d)(?=.*[!@#$%^&*()_+\-=[\]{};':"\\|,.<>\/?]).{8,20}$/
+        /^(?=.*[A-Za-z])(?=.*\d)(?=.*[!@#$%^&*()_+\-=[\]{};':"\\|,.<>\/?]).{8,20}$/,
       )
       .messages({
         "string.pattern.base": "영문, 숫자, 특수문자를 포함해주세요.",
@@ -41,10 +44,13 @@ const RegisterForm = () => {
         "string.min": "비밀번호는 최소 8글자 이상 입력해주세요.",
         "string.max": "비밀번호는 최대 20글자 이하로 입력해주세요.",
       }),
-    confirmPassword: Joi.string().valid(Joi.ref("password")).required().messages({
-      "string.empty": "",
-      "any.only": "비밀번호가 일치하지 않습니다.",
-    }),
+    confirmPassword: Joi.string()
+      .valid(Joi.ref("password"))
+      .required()
+      .messages({
+        "string.empty": "",
+        "any.only": "비밀번호가 일치하지 않습니다.",
+      }),
   });
 
   const handleChange = (e) => {
@@ -63,7 +69,7 @@ const RegisterForm = () => {
         setActive(false);
       } else {
         setErrors({});
-        setActive(emailVerified);
+        setActive(isEmailVerified);
       }
       return updatedData;
     });
@@ -88,9 +94,12 @@ const RegisterForm = () => {
 
   const handleEmailVerification = async () => {
     try {
-      const response = await axios.post("http://localhost:3000/api/v1/email/send", {
-        mail: formData.email,
-      });
+      const response = await axios.post(
+        "http://localhost:3000/api/v1/email/send",
+        {
+          mail: formData.email,
+        },
+      );
 
       console.log("이메일 인증 코드 전송 성공:", response.data);
       alert("이메일로 인증 코드가 전송되었습니다.");
@@ -102,11 +111,14 @@ const RegisterForm = () => {
 
   const handleVerifyCode = async () => {
     try {
-      const response = await axios.post("http://localhost:3000/api/v1/email/verify", {
-        mail: formData.email,
-        verifyCode: verificationCode,
-      });
-  
+      const response = await axios.post(
+        "http://localhost:3000/api/v1/email/verify",
+        {
+          mail: formData.email,
+          verifyCode: verificationCode,
+        },
+      );
+
       console.log("이메일 인증 성공:", response.data);
       setIsEmailVerified(true);
       alert("이메일 인증이 완료되었습니다.");
@@ -150,15 +162,15 @@ const RegisterForm = () => {
               <span className="register-error-message">{errors.email}</span>
             )}
           </div>
-          {!emailVerified && (
+          {/* {!isEmailVerified && (
             <button
               type="button"
-              onClick={sendVerificationEmail}
+              onClick={handleEmailVerification}
               disabled={emailSent}
             >
               {emailSent ? "이메일 전송 완료" : "이메일 인증하기"}
             </button>
-          )}
+          )} */}
           <div className="register-input-wrap input-password">
             <input
               name="password"
@@ -182,12 +194,16 @@ const RegisterForm = () => {
               required
             />
             {errors.confirmPassword && (
-              <span className="register-error-message">{errors.confirmPassword}</span>
+              <span className="register-error-message">
+                {errors.confirmPassword}
+              </span>
             )}
           </div>
         </div>
         <button
-          className={active ? "active-register-submit-btn" : "register-submit-btn"}
+          className={
+            active ? "active-register-submit-btn" : "register-submit-btn"
+          }
           type="submit"
           disabled={!active}
           onClick={handleEmailVerification}
@@ -199,7 +215,9 @@ const RegisterForm = () => {
       {/* 이메일 인증 모달 */}
       {isModalOpen && !isEmailVerified && (
         <div className="email-verification-modal">
-          <button className="modal-close-btn" onClick={handleModalClose}>×</button>
+          <button className="modal-close-btn" onClick={handleModalClose}>
+            ×
+          </button>
           <h3>이메일 인증</h3>
           <p>이메일로 인증 코드를 전송했습니다. 코드를 입력해주세요.</p>
           <input
@@ -209,7 +227,9 @@ const RegisterForm = () => {
             onChange={(e) => setVerificationCode(e.target.value)}
           />
           <button onClick={handleVerifyCode}>인증하기</button>
-          <button onClick={handleEmailVerification}>이메일 인증 코드 재전송</button>
+          <button onClick={handleEmailVerification}>
+            이메일 인증 코드 재전송
+          </button>
         </div>
       )}
     </div>
