+++ src/oauth/KakaoCallback.jsx
@@ -4,11 +4,12 @@ import axios from "axios";
 
 const KakaoCallback = () => {
   const navigate = useNavigate();
+
   useEffect(() => {
     const handleKakaoLogin = async () => {
       const query = new URLSearchParams(window.location.search);
       const code = query.get("code");
-      
+
       if (!code) {
         console.error("Authorization code not found");
         return;
@@ -18,7 +19,7 @@ const KakaoCallback = () => {
         // 서버에 인증 코드로 액세스 토큰 요청
         const response = await axios.post(
           "http://localhost:3000/api/v1/user/callback/kakao",
-          { code }
+          { code },
         );
         if (response.data.data.accessToken) {
           localStorage.setItem("accessToken", response.data.data.accessToken);
