+++ src/oauth/NaverCallback.jsx
@@ -2,10 +2,11 @@ import { useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 import axios from "axios";
 
-const GoogleCallback = () => {
+const NaverCallback = () => {
   const navigate = useNavigate();
+
   useEffect(() => {
-    const handleGoogleLogin = async () => {
+    const handleNaverLogin = async () => {
       const query = new URLSearchParams(window.location.search);
       const code = query.get("code");
 
@@ -17,8 +18,8 @@ const GoogleCallback = () => {
       try {
         // 서버에 인증 코드로 액세스 토큰 요청
         const response = await axios.post(
-            "http://localhost:3000/api/v1/user/callback/google",
-            { code }
+          "http://localhost:3000/api/v1/user/callback/naver",
+          { code },
         );
         if (response.data.data.accessToken) {
           localStorage.setItem("accessToken", response.data.data.accessToken);
@@ -27,15 +28,15 @@ const GoogleCallback = () => {
           localStorage.setItem("username", response.data.data.username);
           navigate("/");
         } else {
-          console.error("구글 로그인 실패:", response.data);
+          console.error("네이버 로그인 실패:", response.data);
         }
       } catch (error) {
-        console.error("구글 로그인 처리 중 오류:", error);
+        console.error("네이버 로그인 처리 중 오류:", error);
       }
     };
 
-    handleGoogleLogin();
+    handleNaverLogin();
   }, []);
 };
 
-export default GoogleCallback;
+export default NaverCallback;
