+++ src/pages/FindPasswordPage.css
@@ -1,4 +1,4 @@
-.find-password {
+/* .find-password {
   max-width: 400px;
   margin: 0 auto;
   padding: 20px;
@@ -45,4 +45,4 @@ button:disabled {
   margin-top: 20px;
   font-size: 20px;
   color: green;
-}
+} */
